# docs/api/openapi/linksee-v1.yaml
openapi: 3.0.3
info:
  title: Linksee Teaching Collaboration API
  version: 1.0.0
  description: >
    Linksee 教学项目协作与过程评价平台的 P0 API schema。
    本文件与 docs/api/contracts/communication-contract-v1.md 保持一致，
    优先覆盖课程项目、阶段要求、小组讨论、阶段提交、教师反馈和课程看板。
servers:
  - url: /api/v1
    description: Linksee API v1
...
paths:
...
  /stages/{stageId}/groups/{groupId}/submissions:
    post:
      tags:
        - Submissions
      summary: 创建阶段成果提交
      operationId: createSubmission
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/StageId"
        - $ref: "#/components/parameters/GroupId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmissionRequest"
            example:
              title: 需求分析阶段提交
              description: 提交需求说明书、用例图和原型链接
              fileIds:
                - file_001
              links:
                - https://example.com/prototype
              repositoryUrl: https://github.com/example/linksee-demo
              contributionNote: 张三负责需求文档，李四负责原型，王五负责接口草案。
          multipart/form-data:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                description:
                  type: string
                links:
                  type: array
                  items:
                    type: string
                    format: uri
                repositoryUrl:
                  type: string
                  format: uri
                contributionNote:
                  type: string
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: 阶段成果提交创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
              example:
                id: submission_001
                stageId: stage_001
                groupId: group_001
                attemptNo: 2
                replacedSubmissionId: submission_000
                status: submitted
                submittedBy: student_001
                submittedAt: "2026-05-15T21:30:00+08:00"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
    get:
      tags:
        - Submissions
      summary: 获取阶段成果提交列表
      operationId: listStageSubmissions
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/StageId"
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: 阶段成果提交列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    StageId:
      name: stageId
      in: path
      required: true
      schema:
        type: string
      description: Stage ID。
    GroupId:
      name: groupId
      in: path
      required: true
      schema:
        type: string
      description: Group ID。
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 64
      description: 写操作幂等键，建议使用 UUID。
  responses:
    Unauthenticated:
      description: 未认证或 token 无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: UNAUTHENTICATED
            message: Authentication token is missing or invalid.
            details: null
            requestId: trace_20260430_001
    Forbidden:
      description: 当前用户无权访问该资源或执行该操作
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: FORBIDDEN
            message: Current user is not allowed to perform this operation.
            details: null
            requestId: trace_20260430_002
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: NOT_FOUND
            message: Requested resource was not found.
            details: null
            requestId: trace_20260430_003
    Conflict:
      description: 当前资源状态与请求操作冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: CONFLICT
            message: Resource state conflicts with this operation.
            details: null
            requestId: trace_20260430_004
    ValidationFailed:
      description: 请求字段校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: VALIDATION_FAILED
            message: Request payload validation failed.
            details: null
            requestId: trace_20260430_005
  schemas:
    ErrorResponse:
      type: object
      required:
        - ok
        - code
        - message
      properties:
        ok:
          type: boolean
          enum: [false]
        code:
          type: string
          enum:
            - UNAUTHENTICATED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - VALIDATION_FAILED
            - INTERNAL_ERROR
            - FORCE_CHANGE_PASSWORD
        message:
          type: string
        details:
          nullable: true
        requestId:
          type: string
          nullable: true
...
    CreateSubmissionRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 3000
        fileIds:
          type: array
          maxItems: 20
          items:
            type: string
        links:
          type: array
          maxItems: 10
          items:
            type: string
            format: uri
        repositoryUrl:
          type: string
          format: uri
        contributionNote:
          type: string
          maxLength: 3000
    SubmissionStatus:
      type: string
      enum:
        - draft
        - not_submitted
        - submitted
        - under_review
        - needs_changes
        - reviewed
        - approved
        - rejected
    Submission:
      type: object
      required:
        - id
        - stageId
        - groupId
        - status
        - createdBy
        - createdAt
      properties:
        id:
          type: string
        stageId:
          type: string
        groupId:
          type: string
        attemptNo:
          type: integer
        replacedSubmissionId:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/SubmissionStatus"
        summary:
          type: string
        payload:
          type: object
        submittedBy:
          type: string
        submittedAt:
          type: string
          format: date-time
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...